Add files via upload
The Automated fish tank system is known to be very useful for feeding , cleaning 
and monitoring activities and that can been utilized in many ways. The idea of 
the Poseidon’s automated fish tank system is originated from a bunch of many 
problems. The feeding process of fish manually doing and measurement of the 
food for specific fishes is one of the main problems in many peoples who are doing
fish farming at home. The second problem is the lack of availability of time to keep 
the aquarium meeting the mandatory fish needs that need to be reviewed regularly 
and the cleaning of water manually 2 weeks or 4 weeks once or every day is 
difficult Along with these there are known that important parameters such as 
condition of pH water and the availability of food is essential to keep the habitat 
conducive for living. Therefore We are implementing the project with automated
Food feeder ,automatic cleaning without human interference ,notification will 
sent to Telegram handle using ESP32Camera Module